# perf: removed unnecessary offset and byteToHex calc