--- a/src/stringify-internal.js
+++ b/src/stringify-internal.js
@@ -1,3 +1,8 @@
+/**
+ * This is the internal `stringify()` function used by the library.
+ * It isn't exposed in the index because it assumes no offset.
+ */
+
 import validate from './validate.js';
 
 /**
@@ -263,35 +268,35 @@ const byteToHex = [
   'ff',
 ];
 
-export function unsafeStringify(arr, offset = 0) {
+export function unsafeStringify(arr) {
   // Note: Be careful editing this code!  It's been tuned for performance
   // and works in ways you may not expect. See https://github.com/uuidjs/uuid/pull/434
   return (
-    byteToHex[arr[offset + 0]] +
-    byteToHex[arr[offset + 1]] +
-    byteToHex[arr[offset + 2]] +
-    byteToHex[arr[offset + 3]] +
+    byteToHex[arr[0]] +
+    byteToHex[arr[1]] +
+    byteToHex[arr[2]] +
+    byteToHex[arr[3]] +
     '-' +
-    byteToHex[arr[offset + 4]] +
-    byteToHex[arr[offset + 5]] +
+    byteToHex[arr[4]] +
+    byteToHex[arr[5]] +
     '-' +
-    byteToHex[arr[offset + 6]] +
-    byteToHex[arr[offset + 7]] +
+    byteToHex[arr[6]] +
+    byteToHex[arr[7]] +
     '-' +
-    byteToHex[arr[offset + 8]] +
-    byteToHex[arr[offset + 9]] +
+    byteToHex[arr[8]] +
+    byteToHex[arr[9]] +
     '-' +
-    byteToHex[arr[offset + 10]] +
-    byteToHex[arr[offset + 11]] +
-    byteToHex[arr[offset + 12]] +
-    byteToHex[arr[offset + 13]] +
-    byteToHex[arr[offset + 14]] +
-    byteToHex[arr[offset + 15]]
+    byteToHex[arr[10]] +
+    byteToHex[arr[11]] +
+    byteToHex[arr[12]] +
+    byteToHex[arr[13]] +
+    byteToHex[arr[14]] +
+    byteToHex[arr[15]]
   ).toLowerCase();
 }
 
-function stringify(arr, offset = 0) {
-  const uuid = unsafeStringify(arr, offset);
+function stringify(arr) {
+  const uuid = unsafeStringify(arr);
   // Consistency check for valid UUID.  If this throws, it's likely due to one
   // of the following:
   // - One or more input array values don't map to a hex octet (leading to