@broofa 2023-01-08
Thanks for the contribution, however I'm going to pass on this PR at this time.  The perf improvement doesn't justify the increased bundle size, or the maintenance burden of having multiple, separate implementations for `stringify()`.