@Cadienvan 2023-01-08
What about just going for the bytesToHex?